Refactor: renaming QwikSpeak component to QwikSpeakProvider
